refator vcf to bq script
